requirements.txt
pytest>=8.2
pytest-html>=4.1
python-dotenv>=1.0
playwright>=1.47
